fix: add checks in routes.py to prevent from foreceful edit and deletion

+++ bookspire/routes.py
@@ -5,12 +5,14 @@ from werkzeug.security import generate_password_hash, check_password_hash
 
 
 @app.route("/")
+# renders template for homepage
 def home():
     books = list(Book.query.order_by(-Book.id).all())
     return render_template("index.html", books=books)
 
 
 @app.route("/sorted")
+# sorts book by their score
 def sorted():
     books = list(Book.query.order_by(-Book.score).all())
     return render_template("index.html", books=books)
@@ -18,51 +20,54 @@ def sorted():
 
 @app.route("/add_book", methods=["GET", "POST"])
 def add_book():
-    if request.method == "POST":
-        book = Book(
-            title=request.form.get("title"),
-            author=request.form.get("author"),
-            year_of_publication=request.form.get("year_of_publication"),
-            synopsis=request.form.get("synopsis"),
-            genre=request.form.get("genre"),
-            username=session["user"],
-            cover=request.form.get("cover"),
-            score=0,
-        )
-        db.session.add(book)
-        db.session.commit()
-        return redirect(url_for("home"))
-    return render_template("add_book.html")
+    # checks if user is logged in if not redirects to homepage
+    if session["user"]:
+        # adds book to database and redirects home
+        if request.method == "POST":
+            book = Book(
+                title=request.form.get("title"),
+                author=request.form.get("author"),
+                year_of_publication=request.form.get("year_of_publication"),
+                synopsis=request.form.get("synopsis"),
+                genre=request.form.get("genre"),
+                username=session["user"],
+                cover=request.form.get("cover"),
+                score=0,
+            )
+            db.session.add(book)
+            db.session.commit()
+            return redirect(url_for("home"))
+        return render_template("add_book.html")
+    return redirect(url_for("home"))
 
 
 @app.route("/register", methods=["GET", "POST"])
 def register():
     if request.method == "POST":
-        # Check  if username already exists
+        # checks if username already exists
         existing_user = User.query.filter(
             User.username == request.form.get("username").lower()).all()
 
-        # If username exists - flash message & reload register page
+        # if username exists - flash message & reload register page
         if existing_user:
             flash("This username already exists, please try another username.")
             return redirect(url_for("register"))
 
-        # If username doesn't exist
+        # if username doesn't exist
         newuser = User(
             username=request.form.get("username").lower(),
             password=generate_password_hash(request.form.get("password")),
         )
-        # Add to database let user know they are registered
+        # add to database let user know they are registered
         db.session.add(newuser)
         db.session.commit()
         flash('Registration successful!')
 
-        # Add the user to the session and redirect to the main page
+        # add the user to the session and redirect to the main page
         session["user"] = request.form.get("username").lower()
         flash("Welcome, {}. You are now logged in!".format(
             request.form.get("username").capitalize()))
         return redirect(url_for("home"))
-
     return render_template("register.html")
 
 
@@ -91,14 +96,13 @@ def login():
                 return redirect(url_for("login"))
 
         # username doesn't exist
-        else:   
+        else:
             flash("Incorrect Username and/or Password")
-
     return render_template("login.html")
 
 
 @app.route("/logout")
-# logs out user and flashed the message and redirect user to homepage
+# logs out user, flashes the message and redirects user to homepage
 def logout():
     session.pop("user")
     flash("You have been logged out.")
@@ -114,49 +118,59 @@ def book(book_id):
 
 
 @app.route("/delete_book/<int:book_id>")
-#deletes book
+# deletes book
 def delete_book(book_id):
     book = Book.query.get_or_404(book_id)
-    db.session.delete(book)
-    db.session.commit()
-    flash('Book has been deleted!')
+    # checks again if user is the author of the book to be deleted
+    # to prevent forcing deletion
+    if book.username == session["user"]:
+        db.session.delete(book)
+        db.session.commit()
+        flash('Book has been deleted!')
+        return redirect(url_for("home"))
     return redirect(url_for("home"))
 
 
 @app.route("/edit_book/<int:book_id>", methods=["GET", "POST"])
 def edit_book(book_id):
     book = Book.query.get_or_404(book_id)
-    if request.method == "POST":
-        book.title = request.form.get("title")
-        book.author = request.form.get("author")
-        book.year_of_publication = request.form.get("year_of_publication")
-        book.synopsis = request.form.get("synopsis")
-        book.genre = request.form.get("genre")
-        book.cover = request.form.get("cover")
-        db.session.commit()
-        return redirect(url_for("home"))
-
-    #renders template with specific book's details loaded in
-    return render_template("edit_book.html", book=book)
+    # checks again if user is the author of the book to be
+    # edited to prevent forcing edit
+    if book.username == session["user"]:
+        if request.method == "POST":
+            book.title = request.form.get("title")
+            book.author = request.form.get("author")
+            book.year_of_publication = request.form.get("year_of_publication")
+            book.synopsis = request.form.get("synopsis")
+            book.genre = request.form.get("genre")
+            book.cover = request.form.get("cover")
+            db.session.commit()
+            return redirect(url_for("home"))
+        # renders template with specific book's details loaded in
+        return render_template("edit_book.html", book=book)
+    return redirect(url_for("home"))
 
 
 @app.route("/add_review/<int:book_id>", methods=["GET", "POST"])
 def add_review(book_id):
     book = Book.query.get_or_404(book_id)
-    if request.method == "POST":
-        review = Review(
-            book_id=book.id,
-            review_title=request.form.get("review_title"),
-            review_text=request.form.get("review_text"),
-            review_score=request.form.get("review_score"),
-            username=session["user"],
-        )
-        book.score = book.score + int(review.review_score)
-        db.session.add(review)
-        db.session.commit()
-        return redirect(url_for("home"))
-        
-    return render_template("add_review.html", book=book)
+    # checks if user is logged in if not redirects to homepage
+    if session["user"]:
+        # adds review to database
+        if request.method == "POST":
+            review = Review(
+                book_id=book.id,
+                review_title=request.form.get("review_title"),
+                review_text=request.form.get("review_text"),
+                review_score=request.form.get("review_score"),
+                username=session["user"],
+            )
+            book.score = book.score + int(review.review_score)
+            db.session.add(review)
+            db.session.commit()
+            return redirect(url_for("home"))  
+        return render_template("add_review.html", book=book)
+    return redirect(url_for("home"))  
 
 
 @app.route("/edit_review/<int:book_id>/<int:review_id>", methods=["GET", "POST"])
@@ -164,16 +178,20 @@ def edit_review(review_id, book_id):
     review = Review.query.get_or_404(review_id)
     book = Book.query.get_or_404(book_id)
     book.score = book.score - int(review.review_score)
-    if request.method == "POST":
-        review.review_title = request.form.get("review_title")
-        review.review_text = request.form.get("review_text")
-        review.review_score = request.form.get("review_score")
-        book.score = book.score + int(review.review_score)
-        db.session.commit()
-        return redirect(url_for("home"))
+    # checks again if user is the author of the review to be
+    # edited to prevent forcing edit
+    if review.username == session["user"]:
+        if request.method == "POST":
+            review.review_title = request.form.get("review_title")
+            review.review_text = request.form.get("review_text")
+            review.review_score = request.form.get("review_score")
+            book.score = book.score + int(review.review_score)
+            db.session.commit()
+            return redirect(url_for("home"))
 
-    # renders template with specific review details form loaded in
-    return render_template("edit_review.html", review=review, book=book)
+        # renders template with specific review details form loaded in
+        return render_template("edit_review.html", review=review, book=book)
+    return redirect(url_for("home"))
 
 
 @app.route("/delete_revew/<int:book_id>/<int:review_id>")
@@ -182,9 +200,13 @@ def delete_review(review_id, book_id):
     review = Review.query.get_or_404(review_id)
     book = Book.query.get_or_404(book_id)
     book.score = book.score - int(review.review_score)
-    db.session.delete(review)
-    db.session.commit()
-    flash('Review has been deleted!')
+    # checks again if user is the author of the review to be
+    # edited to prevent forcing delete
+    if review.username == session["user"]:
+        db.session.delete(review)
+        db.session.commit()
+        flash('Review has been deleted!')
+        return redirect(url_for("home"))
     return redirect(url_for("home"))
